refactor(add-meal): give the Anthropic paths one request and one schema
Pulls the endpoint, headers, forced tool call, response parsing and — most
importantly — the tool schema out of AnthropicMealTextInterpreter into
AnthropicMealItemsApi. The interpreter is now its prompt and nothing else.

The schema is the reason. It is the app's provenance guarantee in
executable form: it has no macro fields, so a model has nowhere to put a
calorie count. That guarantee is only worth having if reviewing it is one
question — "did anyone add a field to this schema?" — rather than one
question per caller. A second path with its own copy would make it two,
and the copies would drift.

Also renames MealTextInterpreterException to MealInterpreterException,
since nothing about it was ever text-specific. The old name is re-exported
from meal_text_interpreter.dart so existing imports are untouched.

No behaviour change: all 34 existing tests pass with only the mechanical
rename applied to them.
